apps/api/requirements.txt
fastapi
uvicorn
pydantic>=2.0
pydantic-settings
sqlmodel
alembic
psycopg[binary]
redis
rq
langgraph
neo4j
sse-starlette
openai>=1.0.0
pytest
pytest-asyncio
httpx
python-dotenv>=1.2.0
ruff
langchain
langchain-community
langchain-openai>=0.1.0
pgvector
python-docx
pypdf
python-multipart
croniter
